Update Day 05 Loops.jl

+++ Practice/Tutorials/30 Days of Code/Day 05 Loops.jl
@@ -1,8 +1,8 @@
 # https://www.hackerrank.com/challenges/30-loops/problem
 
-n = parse(Int64, readline())
+n = parse(Int, readline())
 
-function show_loop(v::Int64)
+function show_loop(v::Int)
     for i in 1:10
         r = v * i
         println("$v x $i = $r")
